fix: skip contract verification on localnet & refactor
